\name{mypackage-internal}
\alias{BIC.means.nop}
\alias{BIC.means}
\alias{Change.Points.R}
\alias{Chisq.Contrib.R}
\alias{Chisq.Contrib.C}
\alias{Chisq.Contrib.test}
\alias{Remove.Overlap.R}
\alias{Remove.Overlap.C}
\alias{Remove.Overlap.test}
\alias{ComputeZ.R}
\alias{computeZ.onechange}
\alias{computeZ.onechange.sample}
\alias{computeZ.squarewave.sample}
\alias{ComputeZ.C}
\alias{computeZ.test}
\alias{flatten}
\alias{plotChpts}
\alias{getCutoffMultisampleWeightedChisq}
\alias{pvalueMultisampleWeightedChisq}
\alias{pvalueMultisampleWeightedChisqold}
\alias{mllr}
\alias{mllrdot}
\alias{wchi}
\alias{wchidot}
\alias{wschi}
\alias{nuFunction}
\alias{psiNormal}
\alias{computeLocalIncrementMean}
\alias{getCutoffMultisample}
\alias{pvalueMultisample}
\alias{pvalueSumChisqNew}
\alias{pvalueSumChisq}
\alias{dchi}
\alias{pmarg.sumweightedchisq}
\alias{computeMomentsBinomial}
\alias{computeMoments}
\alias{computeTiltDirect}
\alias{vu}
\alias{getCutoffEpidemicChangeLocationKnown}
\alias{computeBeta}
\alias{intmode}
\alias{msscan.old}
\alias{msscan.mbic}
\alias{update.bic.baseline}
\alias{mscbs.old}
\alias{mbic.ms}
\alias{next.tau}
\alias{compute.var}
\alias{mscbs.classify}
\alias{compute.max.Z.C}
\alias{feature.matrix.scan}
\alias{feature.matrix.cbs}
\alias{cnvtable}
\alias{norm.snp}
\alias{matrix.max}
\alias{compute.max.Z}
\alias{compute.max.ProjectedZ}
\alias{fcompute.max.ProjectedZ}
\alias{ComputeZ.fromS.anyg.R}
\alias{ComputeZ.fromS.R}
\alias{ComputeZ.fromS.R.partial}
\alias{ComputeBYZ.fromS.R}
\alias{ComputeBYZ.fromS.R.partial}
\alias{ComputeProjectedZ.fromS.R}
\alias{ComputeProjectedZ.fromS.R.partial}
\alias{Chisq.Contrib.fromS.R}
\alias{merge.pos}
\alias{fscan}
\alias{scan.classify}
\alias{scan.classify.seg}
\alias{mscbs.get.fitted}
\alias{mscbs.classify.bic}
\alias{fcompute.max.Z}
\alias{fscan.max}
\alias{scoreFit}


\title{Internal mpcbs Functions}
\description{
    Internal mscp functions
}
\details{
    These are not to be called by the user.
}
\keyword{ internal }

