\name{sqcp}
          \docType{data}
          \alias{sqcp}
          \title{Example copy number data from the 62 normal samples of Stanford Quality Control Panel.}
          \description{
            This data set contains an example copy number data set taken from snps # 1501 to 3000 of chromosome 22 of Stanford Quality Control Panel.   
          }
          \usage{sqcp}
          \format{Rows are locations, columns are samples.
                
          }
          \references{
Zhang, N.R., Siegmund, D.O., Ji, H., and Li, J., 2010, Detecting simultaneous change-points in multiple sequences. Biometrika, 97, 631--645. \cr
    Siegmund, D.O., Yakir, B. and Zhang, N.R., 2011, Detecting simultaneous variant intervals in aligned sequences.  Annals of Applied Statistics, in press. \cr
 
                     }
          \keyword{datasets}
